fix Next route export for PDF build
